+++ src/ios/FirebasePlugin.m
@@ -8,6 +8,7 @@
 @import FirebaseAnalytics;
 @import FirebaseRemoteConfig;
 @import FirebasePerformance;
+@import FirebaseCore;
 @import FirebaseAuth;
 @import FirebaseFunctions;
 @import UserNotifications;
@@ -41,6 +42,7 @@ static NSDictionary* googlePlist;
 static NSMutableDictionary* firestoreListeners;
 static NSString* currentInstallationId;
 static NSMutableDictionary* traces;
+static FIROAuthProvider* oauthProvider;
 static FIRMultiFactorResolver* multiFactorResolver;
 
 
@@ -72,7 +74,7 @@ static FIRMultiFactorResolver* multiFactorResolver;
         googlePlist = [NSMutableDictionary dictionaryWithContentsOfFile:[[NSBundle mainBundle] pathForResource:@"GoogleService-Info" ofType:@"plist"]];
 
         [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(applicationLaunchedWithUrl:) name:CDVPluginHandleOpenURLNotification object:nil];
-
+        
         if([self getGooglePlistFlagWithDefaultValue:FirebaseCrashlyticsCollectionEnabled defaultValue:YES]){
             [self setPreferenceFlag:FIREBASE_CRASHLYTICS_COLLECTION_ENABLED flag:YES];
         }
@@ -1025,6 +1027,30 @@ static FIRMultiFactorResolver* multiFactorResolver;
         }
 
         [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+    }@catch (NSException *exception) {
+        [self handlePluginExceptionWithContext:exception :command];
+    }
+}
+
+- (void)authenticateUserWithMicrosoft:(CDVInvokedUrlCommand*)command{
+    @try {
+        oauthProvider = [FIROAuthProvider providerWithProviderID:@"microsoft.com"];
+        [oauthProvider setCustomParameters:@{@"prompt": @"consent"}];
+        [oauthProvider getCredentialWithUIDelegate:nil
+                            completion:^(FIRAuthCredential *_Nullable credential, NSError *_Nullable error) {
+            if (error) {
+                NSLog(@"Error: %@ %@", error, [error userInfo]);
+                @throw([NSException exceptionWithName:@"Error" reason:error.localizedDescription userInfo:nil]);
+            }
+            if (credential) {
+                NSNumber* key = [self saveAuthCredential:credential];
+                NSMutableDictionary* result = [[NSMutableDictionary alloc] init];
+                [result setValue:@"true" forKey:@"instantVerification"];
+                [result setValue:key forKey:@"id"];
+                CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:result];
+                [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+            }
+        }];
     }@catch (NSException *exception) {
         [self handlePluginExceptionWithContext:exception :command];
     }