+++ types/index.d.ts
@@ -242,11 +242,16 @@ export interface FirebasePlugin {
     ): void
     authenticateUserWithGoogle(
         clientId: string,
-        success?: () => void,
+        success?: (credential:object) => void,
         error?: (err: string) => void
     ): void
     authenticateUserWithApple(
-        success?: () => void,
+        success?: (credential:object) => void,
+        error?: (err: string) => void,
+        locale?: string,
+    ): void
+    authenticateUserWithMicrosoft(
+        success?: (credential:object) => void,
         error?: (err: string) => void,
         locale?: string,
     ): void