+++ www/firebase.js
@@ -339,6 +339,10 @@ exports.authenticateUserWithApple = function (success, error, locale) {
     exec(success, error, "FirebasePlugin", "authenticateUserWithApple", [locale]);
 };
 
+exports.authenticateUserWithMicrosoft = function (success, error, locale) {
+  exec(success, error, "FirebasePlugin", "authenticateUserWithMicrosoft", [locale]);
+};
+
 exports.signInWithCredential = function (credential, success, error) {
     if(typeof credential !== 'object') return error("'credential' must be an object");
     exec(success, handleAuthErrorResult(error), "FirebasePlugin", "signInWithCredential", [credential]);